backend/pyproject.toml
[project]
name = "todo-app-backend"
version = "0.1.0"
description = "Phase II Multi-User Todo Application Backend"
requires-python = ">=3.13"
dependencies = [
    "sqlmodel>=0.0.14",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
